clean up main program source
